moved all important scripts to Scripts folder
- Added code level documentation
